Home230402

+++ src/je/panse/doro/commoncode/item_administratus/file/ChartplateCreator.java
@@ -1,6 +1,6 @@
 package je.panse.doro.commoncode.item_administratus.file;
 
-import java.io.*;			
+import java.io.*;					
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
@@ -12,8 +12,14 @@ public class ChartplateCreator {
     String currentUsersDir = System.getProperty("user.dir");
     String homeDir = currentUsersDir + "/src";
     String workingDir = homeDir + "/datatext/samsara";
-    List<String> subsourceFiles = Arrays.asList("3CC", "3PI", "3PMH", "3ROS", "4PMH", "5SUJ", "6OBJ", "7LAB", "8ASS", "9PLAN", "10COMMENT");
+    List<String> subsourceFiles = Arrays.asList("3CC", "3PI", "3ROS", "4PMH", "5SUJ", "6OBJ", "7LAB", "8ASS", "9PLAN", "10COMMENT");
     File chartplateFile = new File(EntryDir.workingDir + "/chartplate"); // Change the file name and path as per your requirements
+    
+    System.out.println(chartplateFile);
+    if (!chartplateFile.exists() || chartplateFile.listFiles() == null || chartplateFile.listFiles().length == 0) {
+        System.out.println("Error: chartplate directory is empty or does not exist.");
+        return;
+    }
     try {
       FileWriter writer = new FileWriter(chartplateFile);
       BufferedWriter bufferWriter = new BufferedWriter(writer);
